# swagger.yaml
swagger: '2.0'
basePath: /api/v0.1.1
info:
  title: WP3-backed Data Service API
  description: An API which supports querying available LEXIS data sets
  version: 0.1.1
  contact:
    email: ruben@lrz.de
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
definitions:
...
  RequestID:
    type: object
    required:
      - request_id
    properties:
      request_id:
        type: string
        format: uuid
        example: "31279ed6-e7b1-4c9f-ab26-206c863a2264"
        description: UUID
...
  ErrorResponse:
    type: object
    required:
      - errorString
    properties:
      errorString:
        type: string
...
  DataSize:
    type: object
    required:
      - result
    properties:
      result:
        type: string
        description: Status of the datasize get process
        example: "Done"
      size:
        type: string
        description: dataset size
        example: "45645352354"
      totalfiles:
        type: string
        description: dataset files
        example: "343"
      smallfiles:
        type: string
        description: dataset files smaller than 32 MB
        example: "2"
...
securityDefinitions:
  Keycloak:
    type: oauth2
    flow: accessCode
    authorizationUrl: 'https:138.246.232.245:6443/auth'
    tokenUrl: 'https:138.246.232.245:6443/auth'
security:
  - Keycloak: []
paths:
...
  /data/size:
    post:
      tags:
        - dataSetManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: This is called when a user requests size of a dataset.
      operationId: Size
      description: This is called when a user requests size of a dataset.
      parameters:
        - in: body
          name: parameters
          description: parameters
          required: true
          schema:
            type: object
            required:
              - target_system
              - target_path
            properties:
              target_system:
                type: string
              target_path:
                type: string
      responses:
        '201':
          description: This is called when a user requests size of a dataset.
          schema:
            $ref: "#/definitions/RequestID"
        '400':
          description: This means that there's something wrong in the input parameters and the server couldn't understand the request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: This means that the resource the user is trying to transfer from or to is not readable or writable by the user. User doesn't have the correct rights to either read the source file or write on the target system location.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description:  This means that either the source path or the the target path on the system doesn't exist.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '414':
          description: This means that the either the source path or the target path is longer than the server is willing to interpret.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '429':
          description: This means that the user has sent too many requests in a given amount of time. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: This means that something has gone wrong on the data transfer steering server. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /data/size/{request_id}:
    get:
      tags:
        - dataSetManagement
      produces:
        - application/json
      summary: This is called when a user requests to check the status of datasize request
      operationId: CheckSizeStatus
      description: This is called when a user requests to check the status of datasize request
      parameters:
        - in: path
          name: request_id
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: This means that the status has been returned to the user in the response body.
          schema:
            $ref: "#/definitions/DataSize"
        '400':
          description: This means that there's something wrong in the input parameters and the server couldn't understand the request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: This means that the resource the user is trying to transfer from or to is not readable or writable by the user. User doesn't have the correct rights to either read the source file or write on the target system location.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description:  This means that either the source path or the the target path on the system doesn't exist.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '414':
          description: This means that the either the source path or the target path is longer than the server is willing to interpret.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '429':
          description: This means that the user has sent too many requests in a given amount of time. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: This means that something has gone wrong on the data transfer steering server. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
...